RAG/requirements.txt
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
